initial backend implementation - working
