+++ src/history_book/data_models/kg_entities.py
@@ -1,0 +1,67 @@
+"""Data models for knowledge graph entities stored in Weaviate."""
+
+import uuid
+from datetime import UTC, datetime
+from typing import ClassVar
+
+from pydantic import BaseModel, Field
+
+
+class KGEntity(BaseModel):
+    """A normalized entity in a knowledge graph.
+
+    Stored in Weaviate with vector search on search_text.
+    """
+
+    id: str | None = Field(default_factory=lambda: str(uuid.uuid4()))
+    graph_name: str
+    name: str
+    entity_type: str  # person, polity, place, event
+    aliases: list[str] = Field(default_factory=list)
+    description: str = ""
+    occurrence_count: int = 1
+    book_indices: list[int] = Field(default_factory=list)
+    source_book_chapters: list[str] = Field(
+        default_factory=list
+    )  # ["3:2", "3:3"] — book:chapter provenance
+    source_pages: list[int] = Field(default_factory=list)
+    source_paragraph_ids: list[str] = Field(default_factory=list)
+    merged_from_count: int = 0
+    search_text: str = ""  # "Name: X | Type: Y | Description: Z | Aliases: ..."
+
+    vectorize_fields: ClassVar[list[str]] = ["search_text"]
+
+
+class KGRelationship(BaseModel):
+    """A relationship between two entities in a knowledge graph."""
+
+    id: str | None = Field(default_factory=lambda: str(uuid.uuid4()))
+    graph_name: str
+    source_entity_id: str
+    target_entity_id: str
+    entity_ids: list[str] = Field(
+        default_factory=list
+    )  # [source_entity_id, target_entity_id] for ContainsAny queries
+    source_entity_name: str
+    target_entity_name: str
+    relation_type: str  # ruled, conquered, fought, etc.
+    temporal_context: str = ""
+    paragraph_id: str
+    book_index: int
+    chapter_index: int
+    page: int
+
+
+class KGGraph(BaseModel):
+    """Metadata for a named knowledge graph."""
+
+    id: str | None = Field(default_factory=lambda: str(uuid.uuid4()))
+    name: str  # "book3_ch2", "book3", "volume_full"
+    graph_type: str  # "chapter", "book", "volume"
+    book_chapters: list[str] = Field(
+        default_factory=list
+    )  # ["3:0", "3:1", "4:0"] — what's included
+    entity_count: int = 0
+    relationship_count: int = 0
+    created_at: datetime = Field(default_factory=lambda: datetime.now(UTC))
+    updated_at: datetime = Field(default_factory=lambda: datetime.now(UTC))
